requirements.txt
plotly>=5.0.0
pandas>=1.0.0
numpy>=1.0.0
matplotlib>=3.0.0
seaborn>=0.11.0
Pillow>=8.0.0
streamlit>=1.0.0